+++ src/pydocstyle/parser.py
@@ -81,6 +81,13 @@ class Definition(Value):
         return chain([self], *self.children)
 
     @property
+    def error_lineno(self):
+        """Get the line number with which to report violations."""
+        if isinstance(self.docstring, Docstring):
+            return self.docstring.start
+        return self.start
+
+    @property
     def _publicity(self):
         return {True: 'public', False: 'private'}[self.is_public]
 
@@ -210,6 +217,21 @@ class Decorator(Value):
     _fields = 'name arguments'.split()
 
 
+class Docstring(str):
+    """Represent a docstring.
+
+    This is a string, but has additional start/end attributes representing
+    the start and end of the token.
+
+    """
+    def __new__(cls, v, start, end):
+        return str.__new__(cls, v)
+
+    def __init__(self, v, start, end):
+        self.start = start
+        self.end = end
+
+
 VARIADIC_MAGIC_METHODS = ('__init__', '__call__', '__new__')
 
 
@@ -334,7 +356,11 @@ class Parser(object):
             self.log.debug("parsing docstring, token is %r (%s)",
                            self.current.kind, self.current.value)
         if self.current.kind == tk.STRING:
-            docstring = self.current.value
+            docstring = Docstring(
+                self.current.value,
+                self.current.start[0],
+                self.current.end[0]
+            )
             self.stream.move()
             return docstring
         return None